+++ webServer/src/main/java/top/jayu/oa/service/OaBillService.java
@@ -276,7 +276,7 @@ public class OaBillService {
         return candidateLevel;
     }
 
-    private float getOriginNextApproveId(float candidateLevel, Byte billType, Byte applyOrgYesOffice){
+    private float getOriginNextApproveId(float candidateLevel, Integer billType, Byte applyOrgYesOffice){
         OaProcess process = new OaProcess();
         process.setBillType(billType);
         process.setOrgPrivLen(candidateLevel);
@@ -302,7 +302,7 @@ public class OaBillService {
     }
 
     // 获取流程步骤
-    private OaProcess getProcess(Byte billType, String stepName){
+    private OaProcess getProcess(Integer billType, String stepName){
         return oaProcessService.getProcess(billType, stepName);
     }
